services/face-recognition/requirements.txt
flask==2.3.2
flask-cors==3.0.10
numpy==1.24.3
opencv-python==4.7.0.72
onnxruntime-gpu==1.15.1
pillow==9.5.0
scipy==1.10.1
